restructure project to make it deployable again
