--- library/pom.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>org.zalando</groupId>
-	<artifactId>failsafe-actuator</artifactId>
-	<version>0.5.2</version>
-	<packaging>jar</packaging>
-
-	<name>Failsafe-Actuator-Library</name>
-	<description>Failsafe-Actuator provides an out of the box endpoint for Failsafe</description>
-
-	<parent>
-		<groupId>org.zalando</groupId>
-		<artifactId>failsafe-actuator-parent</artifactId>
-		<version>0.5.0</version>
-	</parent>
-
-	<properties>
-		<maven.deploy.skip>false</maven.deploy.skip>
-	</properties>
-    
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-javadoc-plugin</artifactId>
-				<version>2.10.4</version>
-				<executions>
-					<execution>
-						<id>attach-javadocs</id>
-						<goals>
-							<goal>jar</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-gpg-plugin</artifactId>
-				<version>1.6</version>
-				<executions>
-					<execution>
-						<id>sign-artifacts</id>
-						<phase>verify</phase>
-						<goals>
-							<goal>sign</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
-</project>
