+++ src/main/java/org/zalando/failsafeactuator/config/FailsafeAutoConfiguration.java
