+++ src/main/java/org/zalando/failsafeactuator/config/FailsafeInjectionConfiguration.java
