+++ src/main/java/org/zalando/failsafeactuator/endpoint/domain/CircuitBreakerState.java
