+++ src/test/java/org/zalando/failsafeactuator/config/FailsafeAutoConfigurationTest.java
