+++ src/test/java/org/zalando/failsafeactuator/endpoint/FailsafeEndpointTest.java
