+++ src/test/java/org/zalando/failsafeactuator/service/CircuitBreakerRegistryTest.java
